pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lightsail"
version = "0.1.0"
description = "SiN lightsail photonic structure optimization"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
mobo = [
    "torch>=1.13",
    "gpytorch>=1.11",
    "botorch>=0.9",
]
rcwa = [
    "grcwa>=0.1.2",
    "autograd>=1.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
